requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.39.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
Eel==0.16.0
et-xmlfile==1.1.0
flake8==6.1.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.0rc3
isort==5.12.0
mccabe==0.7.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pefile==2023.2.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
pywinstyles==1.4
six==1.16.0
tzdata==2023.3
whichcraft==0.6.1
XlsxWriter==3.1.4
zope.event==5.0
zope.interface==6.0
